feat(auth): Add role-based authorization
We now have a role-base mechanism to access pages of our app. We
have 3 roles:

  * Admin
  * Manager
  * User

Some pages e.g. `home` are public and can be accessed with or without
a session. Others like the `server`, `client` and `about` pages need
a session and then each page has finer grained role-based access.

  * server page - needs a session only
  * about page  - needs a `admin` role
  * client page - needs a `manager` role

The fine-grained role based access is defined in the `middleware.ts`
file, this saves repeating code in each page. If the user doesn't have
the correct rol, then they are shown the `/denied` page instead.